tweet-classifier/requirements.txt
pika>=1.2.0
tweet-preprocessor>=0.6.0